+++ circuits/src/field/decomposition/pow2range.rs
@@ -193,13 +193,12 @@ impl<F: CircuitField> Pow2RangeChip<F> {
         let t_tag = meta.lookup_table_column();
         let t_val = meta.lookup_table_column();
 
-        meta.batched_lookup("pow2range column check", |meta| {
-            let sel = meta.query_selector(q_pow2range);
+        meta.batched_lookup("pow2range column check", Some(q_pow2range), |meta| {
             let tag = meta.query_fixed(tag_col, Rotation::cur());
             let tags = vec![tag; val_cols.len()];
             let vals = val_cols
                 .iter()
-                .map(|val_col| sel.clone() * meta.query_advice(*val_col, Rotation::cur()))
+                .map(|val_col| meta.query_advice(*val_col, Rotation::cur()))
                 .collect();
             vec![(tags, t_tag), (vals, t_val)]
         });