+++ circuits/src/hash/ripemd160/ripemd160_chip.rs
@@ -177,7 +177,7 @@ impl<F: CircuitField> ComposableChip<F> for RipeMD160Chip<F> {
         let q_mod_add = meta.selector();
 
         (0..2).for_each(|idx| {
-            meta.lookup("plain-spreaded lookup", |meta| {
+            meta.lookup("plain-spreaded lookup", None, |meta| {
                 let q_lookup = meta.query_selector(q_lookup);
 
                 let nbits = meta.query_fixed(fixed_cols[idx], Rotation(0));