+++ circuits/src/parsing/automaton_chip.rs
@@ -215,17 +215,16 @@ where
         // is filled).
         let automata = NativeAutomaton::<F>::from_collection(automata);
 
-        meta.lookup("automaton transition check", |meta| {
-            let q = meta.query_selector(q_automaton);
+        meta.lookup("automaton transition check", Some(q_automaton), |meta| {
             let source = meta.query_advice(state_col, Rotation::cur());
             let letter = meta.query_advice(letter_col, Rotation::cur());
             let target = meta.query_advice(state_col, Rotation::next());
             let output = meta.query_advice(output_col, Rotation::cur());
             vec![
-                (q.clone() * source, t_source),
-                (q.clone() * letter, t_letter),
-                (q.clone() * target, t_target),
-                (q * output, t_output),
+                (source, t_source),
+                (letter, t_letter),
+                (target, t_target),
+                (output, t_output),
             ]
         });