+++ circuits/src/verifier/verifier_gadget.rs
@@ -627,6 +627,7 @@ impl<S: SelfEmulation> VerifierGadget<S> {
                     layouter,
                     &self.scalar_chip,
                     &p.evaluated,
+                    argument.selector_expression(),
                     argument.input_expressions(),
                     argument.table_expressions(),
                     &advice_evals,