+++ proofs/examples/proof-size.rs
@@ -34,7 +34,7 @@ impl Circuit<Scalar> for TestCircuit {
             advice: meta.advice_column(),
         };
 
-        meta.lookup("lookup", |meta| {
+        meta.lookup("lookup", None, |meta| {
             let selector = meta.query_selector(config.selector);
             let not_selector = Expression::from(1) - selector.clone();
             let advice = meta.query_advice(config.advice, Rotation::cur());